requirements.txt
numpy
torch
albumentations
streamlit
Pillow